Fix: Move to Cache namespace
