+++ src/Cache/FileCacheManagerInterface.php
@@ -10,7 +10,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace PhpCsFixer;
+namespace PhpCsFixer\Cache;
 
 /**
  * Class supports caching information about state of fixing files.
